+++ src/main/java/com/invicta/lms/repository/RoleRepository.java
@@ -4,7 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.invicta.lms.entity.Role;
 
-public interface RoleRepository extends JpaRepository<Role, Integer> {
-	Role findRoleById(Integer id);
+public interface RoleRepository extends JpaRepository<Role, Long> {
+	Role findRoleById(Long id);
+	
 		
 }
