+++ src/main/java/com/invicta/lms/service/RoleService.java
@@ -7,8 +7,8 @@ import com.invicta.lms.entity.Role;
 public interface RoleService {
 	Role addRole(Role role);
 	List<Role> viewAllRole();
-	Integer deleteRole(Integer id);
-	Role updateRole(Integer id,Role role);
-	Role findRoleById(Integer id);
+	Long deleteRole(Long id);
+	Role updateRole(Long id,Role role);
+	Role findRoleById(Long id);
 	
 }
